// frontend/ctrlvalue/src/app/services/api.generated.ts
export enum AccountType {
    ASSET = "ASSET",
    LIABILITY = "LIABILITY",
}
// ...
export enum AssetClass {
    CASH = "CASH",
    STOCK = "STOCK",
    ETF = "ETF",
    METAL = "METAL",
    VEHICLE = "VEHICLE",
    PROPERTY = "PROPERTY",
    SUPER = "SUPER",
    BUSINESS = "BUSINESS",
    CRYPTO = "CRYPTO",
    OTHER = "OTHER",
}
// ...
export enum LiquidityClass {
    LIQUID = "LIQUID",
    SEMI_LIQUID = "SEMI_LIQUID",
    ILLIQUID = "ILLIQUID",
    LOCKED = "LOCKED",
}